requirements.txt
python-telegram-bot>=20.0
instagrapi
yt-dlp
requests
python-dotenv
psutil
static-ffmpeg
pillow
